@@ -1,6 +1,6 @@
 {
   "name": "speech-sample",
-  "version": "0.5.22",
+  "version": "0.5.23",
   "description": "Speech-Sample",
   "author": "LinguaLogic Team",
   "license": "MIT",
@@ -27,56 +27,55 @@
     "electron": "gulp electron-run"
   },
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/http": "^7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "angular-popper": "^1.0.0",
-    "bootstrap": "^4.1.3",
-    "core-js": "^2.5.7",
-    "jquery": "^3.2.1",
-    "popper.js": "^1.14.5",
-    "rxjs": "^6.3.3",
-    "speech-framework": "file:speech-framework-0.5.22.tgz",
-    "speech-angular": "file:speech-angular-0.5.22.tgz",
-    "tether": "^1.4.5",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.26"
+    "@angular/animations": "~12.1.1",
+    "@angular/common": "~12.1.1",
+    "@angular/compiler": "~12.1.1",
+    "@angular/core": "~12.1.1",
+    "@angular/forms": "~12.1.1",
+    "@angular/platform-browser": "~12.1.1",
+    "@angular/platform-browser-dynamic": "~12.1.1",
+    "@angular/router": "~12.1.1",
+    "@ngx-translate/core": "^13.0.0",
+    "bootstrap": "^5.0.2",
+    "rxjs": "^7.2.0",
+    "speech-angular": "file:speech-angular-0.5.23.tgz",
+    "speech-framework": "file:speech-framework-0.5.23.tgz",
+    "tslib": "^2.1.0",
+    "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.0",
-    "@angular/cli": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@compodoc/compodoc": "^1.1.7",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "~8.9.4",    
+    "@angular-devkit/build-angular": "~12.1.1",
+    "@angular/cli": "~12.1.1",
+    "@angular/compiler-cli": "~12.1.1",
+    "@angular/language-service": "~12.1.1",
+    "@angular/localize": "^12.1.1",
+    "@compodoc/compodoc": "^1.1.13",
+    "@types/chrome": "0.0.146",
+    "@types/jasmine": "~3.8.1",
+    "@types/node": "~16.3.1",
     "child_process": "^1.0.2",
-    "codelyzer": "^4.5.0",
-    "electron-packager": "^13.1.0",
+    "codelyzer": "^6.0.2",
+    "electron-packager": "^15.2.0",
     "gulp": "^4.0.2",
     "gulp-file": "^0.4.0",
-    "gulp-inject-string": "^1.1.1",
-    "gulp-rename": "^1.4.0",
+    "gulp-gzip": "^1.4.2",
+    "gulp-inject-string": "^1.1.2",
+    "gulp-rename": "^2.0.0",
+    "gulp-replace": "^1.1.3",
     "gulp-shell": "^0.8.0",
+    "gulp-tar": "^3.1.0",
     "gulp4-run-sequence": "^1.0.1",
-    "jasmine-core": "~3.3.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.1",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.1",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "3.2.2"
+    "jasmine-core": "~3.8.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.3.4",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-jasmine": "^4.0.1",
+    "karma-jasmine-html-reporter": "^1.6.0",
+    "protractor": "^7.0.0",
+    "ts-node": "~10.0.0",
+    "tslint": "^6.1.3",
+    "typescript": "4.2.4"
   }
 
 }
