@@ -7,8 +7,8 @@ describe('speech-sample App', () => {
     page = new AppPage();
   });
 
-  it('should display welcome message', () => {
+  it('should display welcome message', async () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('Eine Website oder Web-App erweitert um Sprachdienste');
+    expect( await page.getParagraphText()).toEqual('Eine Website oder Web-App erweitert um Sprachdienste');
   });
 });
