@@ -11,10 +11,10 @@
  */
 
 
-export const SPEECHSAMPLE_VERSION_NUMBER = '0.5.22';
-export const SPEECHSAMPLE_VERSION_BUILD = '0023';
+export const SPEECHSAMPLE_VERSION_NUMBER = '0.5.23';
+export const SPEECHSAMPLE_VERSION_BUILD = '0024';
 export const SPEECHSAMPLE_VERSION_TYPE = 'release';
-export const SPEECHSAMPLE_VERSION_DATE = '31.08.2020';
+export const SPEECHSAMPLE_VERSION_DATE = '10.09.2021';
 // tslint:disable-next-line
 export const SPEECHSAMPLE_VERSION_STRING = SPEECHSAMPLE_VERSION_NUMBER + '.' + SPEECHSAMPLE_VERSION_BUILD + ' vom ' + SPEECHSAMPLE_VERSION_DATE + ' (' + SPEECHSAMPLE_VERSION_TYPE + ')';
 
